# nars-api/Controllers/DraftFeaturesController.cs
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.RateLimiting;
using System.Text.Json;
using NarsApi.DTOs;
using NarsApi.Infrastructure;
using NarsApi.Models;
using NarsApi.Services;

namespace NarsApi.Controllers;

[ApiController]
[Route("api/draft-features")]
[Authorize]
public sealed class DraftFeaturesController(
    IDraftFeaturesService draftFeaturesService,
    IRoadGenerationService roadGenerationService,
    ILogger<DraftFeaturesController> logger,
    IWebHostEnvironment webHost) : NarsControllerBase(webHost)
{
    // Accepted image mime types for geo-referenced tiles submitted to segmentation.
    private static readonly HashSet<string> AllowedTileContentTypes = new(StringComparer.OrdinalIgnoreCase)
    {
        "image/tiff", "image/tif", "image/jpeg", "image/jpg", "image/png", "image/webp",
    };

    // File extensions mapped from the accepted mime types above.
    private static readonly HashSet<string> AllowedTileExtensions = new(StringComparer.OrdinalIgnoreCase)
    {
        ".tif", ".tiff", ".jpeg", ".jpg", ".png", ".webp",
    };

    /// <summary>
    /// Submits an imagery tile for the given commune, runs AI
    /// segmentation, and stores the results as pending draft features.
    /// The `featureType` form field ("building" or "road") selects which
    /// model/endpoint to call (per-request separation — each response
    /// carries only one feature type). The caller must have access to the
    /// commune. Does not touch production feature tables.
    /// </summary>
    [HttpPost("segment")]
    [EnableRateLimiting(RateLimitPolicies.Segmentation)]
    // 50MB cap: a 1024x1024 georeferenced tile is typically a few MB, so this
    // allows headroom without disabling limits. Above Sonar's 8MB default
    // threshold by design (see FeatureDefaults:MultipartBodyLengthLimit). The
    // uploaded file's content type and extension are validated below so the
    // large bound is not an arbitrary-upload vector.
#pragma warning disable S5693 // RequestSizeLimit(50MB) is an intentional, bounded upload cap for imagery tiles
    [RequestSizeLimit(50_000_000)]
    [RequestFormLimits(MultipartBodyLengthLimit = 50_000_000)]
#pragma warning restore S5693
    public async Task<ActionResult<SegmentSummaryResponse>> SegmentTile(
        [FromForm] SegmentTileRequest request,
        CancellationToken cancellationToken)
    {
        if (request.CommuneId is null)
        {
            return Problem(detail: "communeId is required.", statusCode: 400);
        }

        if (request.Tile.Length == 0)
        {
            return Problem(detail: "Uploaded tile is empty.", statusCode: 400);
        }

        // Normalize and validate the requested feature type. Default to
        // "building" so existing clients (which omit the field) keep working.
        var featureType = string.IsNullOrWhiteSpace(request.FeatureType)
            ? AiDraftFeature.TypeBuilding
            : request.FeatureType.Trim().ToLowerInvariant();
        if (featureType is not (AiDraftFeature.TypeBuilding or AiDraftFeature.TypeRoad))
        {
            return Problem(detail: "featureType must be 'building' or 'road'.", statusCode: 400);
        }

        var contentType = request.Tile.ContentType;
        var extension = Path.GetExtension(request.Tile.FileName);
        if (!AllowedTileContentTypes.Contains(contentType))
        {
            logger.LogWarning("Rejected tile upload with disallowed content type '{ContentType}'", contentType);
            return Problem(
                detail: "Uploaded tile must be a recognized image type (TIFF, JPEG, PNG or WebP).",
                statusCode: 400);
        }

        if (string.IsNullOrEmpty(extension) || !AllowedTileExtensions.Contains(extension))
        {
            logger.LogWarning("Rejected tile upload with no/unsupported file extension '{FileName}'", request.Tile.FileName);
            return Problem(
                detail: "Uploaded tile must have a recognized image file extension (.tif, .tiff, .jpg, .jpeg, .png or .webp).",
                statusCode: 400);
        }

        if (request.MinLon is null || request.MinLat is null || request.MaxLon is null || request.MaxLat is null)
        {
            return Problem(detail: "minLon, minLat, maxLon and maxLat are required.", statusCode: 400);
        }

        try
        {
            // Ownership of the upload stream transfers to the service: the
            // segmentation client disposes it (through StreamContent) once the
            // request has been dispatched, and the service releases it on its
            // own early-error paths. The controller must not dispose it here.
            var stream = request.Tile.OpenReadStream();
            var summary = await draftFeaturesService.SegmentTileAsync(
                CurrentUserRole,
                CurrentCommuneId,
                CurrentDairaId,
                CurrentWilayaId,
                request.CommuneId.Value,
                featureType,
                stream,
                request.Tile.FileName,
                request.Tile.ContentType,
                (request.MinLon.Value, request.MinLat.Value, request.MaxLon.Value, request.MaxLat.Value),
                cancellationToken);
            return Ok(summary);
        }
        catch (UnauthorizedAccessException)
        {
            return Forbid();
        }
        catch (KeyNotFoundException ex)
        {
            logger.LogWarning(ex, "Segmentation target commune not found: {Reason}", ex.Message);
            return Problem(detail: "The requested commune was not found.", statusCode: 404);
        }
        catch (SegmentationServiceException ex)
        {
            logger.LogError(ex, "Segmentation service call failed for commune {CommuneId}", request.CommuneId);
            return Problem(detail: "Segmentation service is unavailable.", statusCode: 502);
        }
    }

    /// <summary>
    /// Lists draft features for a commune, for the review queue UI.
    /// The caller must have access to the commune.
    /// </summary>
    // Allowlist for draft filter values — rejects unknown values early instead
    // of silently returning empty results.
    private static readonly HashSet<string> AllowedDraftStatuses = new(StringComparer.OrdinalIgnoreCase)
    {
        AiDraftFeature.StatusPending, AiDraftFeature.StatusAccepted,
        AiDraftFeature.StatusRejected, AiDraftFeature.StatusEdited,
    };

    private static readonly HashSet<string> AllowedDraftFeatureTypes = new(StringComparer.OrdinalIgnoreCase)
    {
        AiDraftFeature.TypeRoad, AiDraftFeature.TypeBuilding,
    };

    [HttpGet]
    public async Task<ActionResult<PagedResponse<AiDraftFeatureDto>>> ListDrafts(
        [FromQuery] int communeId,
        [FromQuery] string? featureType,
        [FromQuery] string status = "pending",
        [FromQuery] int skip = 0,
        [FromQuery] int take = 100,
        CancellationToken cancellationToken = default)
    {
        (skip, take) = Pagination.Clamp(skip, take);

        if (!AllowedDraftStatuses.Contains(status))
        {
            var allowed = string.Join(", ", AllowedDraftStatuses.Select(s => $"'{s}'"));
            return Problem(detail: $"status must be one of: {allowed}.", statusCode: 400);
        }

        if (featureType is not null && !AllowedDraftFeatureTypes.Contains(featureType))
        {
            var allowed = string.Join(", ", AllowedDraftFeatureTypes.Select(t => $"'{t}'"));
            return Problem(detail: $"featureType must be one of: {allowed}.", statusCode: 400);
        }

        try
        {
            var drafts = await draftFeaturesService.ListDraftsAsync(
                CurrentUserRole,
                CurrentCommuneId,
                CurrentDairaId,
                CurrentWilayaId,
                communeId,
                featureType,
                status,
                skip,
                take,
                cancellationToken);
            return Ok(drafts);
        }
        catch (UnauthorizedAccessException)
        {
            return Forbid();
        }
    }

    /// <summary>
    /// Accepts a draft: marks it accepted (promotion into the production feature
    /// table is the caller's responsibility — pass GeometryGeoJson/FeatureType
    /// through the same service that handles field-worker-drawn features).
    /// Requires reviewer-level authorization and access to the draft's commune.
    /// </summary>
    [HttpPost("{id:guid}/accept")]
    [Authorize(Policy = "CanReviewFeatures")]
    public async Task<IActionResult> AcceptDraft(Guid id, CancellationToken cancellationToken)
        => await ReviewDraftAsync(id, accept: true, cancellationToken);

    [HttpPost("{id:guid}/reject")]
    [Authorize(Policy = "CanReviewFeatures")]
    public async Task<IActionResult> RejectDraft(Guid id, CancellationToken cancellationToken)
        => await ReviewDraftAsync(id, accept: false, cancellationToken);

    /// <summary>
    /// Edits a pending draft's geometry (rerouting a bad road centerline,
    /// trimming a building footprint) before it is accepted. Requires the same
    /// reviewer scope as accept/reject; an already-reviewed draft is immutable.
    /// </summary>
    [HttpPut("{id:guid}")]
    [Authorize(Policy = "CanReviewFeatures")]
    public async Task<IActionResult> UpdateDraft(Guid id, [FromBody] DraftUpdateRequest body, CancellationToken cancellationToken = default)
    {
        if (!ModelState.IsValid)
        {
            return BadRequest(ModelState);
        }

        return ReviewResultToActionResult(await draftFeaturesService.UpdateDraftAsync(
            CurrentUserRole, CurrentCommuneId, CurrentDairaId, CurrentWilayaId,
            RequiredCurrentUserId, id, body.GeometryGeoJson.Trim(), cancellationToken));
    }

    /// <summary>Deletes a pending draft from the review queue.</summary>
    [HttpDelete("{id:guid}")]
    [Authorize(Policy = "CanReviewFeatures")]
    public async Task<IActionResult> DeleteDraft(Guid id, CancellationToken cancellationToken = default)
        => ReviewResultToActionResult(await draftFeaturesService.DeleteDraftAsync(
            CurrentUserRole, CurrentCommuneId, CurrentDairaId, CurrentWilayaId,
            RequiredCurrentUserId, id, cancellationToken));

    /// <summary>
    /// Materializes pending AI road drafts for a commune into production road
    /// features, enforcing the roads-phase cadastre rules (containment in the
    /// urban areas, turn angle ≤ the configured limit, minimum length/
    /// confidence, connectivity to the existing road network with endpoint
    /// snapping). Drafts that fail a rule stay pending and are reported in
    /// <c>dropped</c>. The caller must have access to the commune, which also
    /// owns the materialized roads.
    /// </summary>
    [HttpPost("generate-roads")]
    public async Task<ActionResult<GenerateRoadsResponse>> GenerateRoads(
        [FromBody] GenerateRoadsRequest request,
        CancellationToken cancellationToken = default)
    {
        if (!ModelState.IsValid)
        {
            return BadRequest(ModelState);
        }

        if (request.CommuneId is null)
        {
            return Problem(detail: "communeId is required.", statusCode: 400);
        }

        try
        {
            var summary = await roadGenerationService.GenerateAsync(
                CurrentUserRole, CurrentCommuneId, CurrentDairaId, CurrentWilayaId,
                RequiredCurrentUserId, request.CommuneId.Value, request.DraftIds, cancellationToken);
            return Ok(new GenerateRoadsResponse(
                summary.Dropped,
                summary.Created
                    .Select(road => new GeneratedRoadDto(
                        road.DbId, road.Layer, road.Label, JsonSerializer.SerializeToElement(road.Data)))
                    .ToList()));
        }
        catch (UnauthorizedAccessException)
        {
            return Forbid();
        }
    }

    private async Task<IActionResult> ReviewDraftAsync(Guid id, bool accept, CancellationToken ct)
    {
        var result = accept
            ? await draftFeaturesService.AcceptDraftAsync(
                CurrentUserRole, CurrentCommuneId, CurrentDairaId, CurrentWilayaId,
                RequiredCurrentUserId, id, ct)
            : await draftFeaturesService.RejectDraftAsync(
                CurrentUserRole, CurrentCommuneId, CurrentDairaId, CurrentWilayaId,
                RequiredCurrentUserId, id, ct);

        return ReviewResultToActionResult(result);
    }

    private IActionResult ReviewResultToActionResult(DraftReviewResult result) => result.Status switch
    {
        DraftReviewStatus.Success => NoContent(),
        DraftReviewStatus.NotFound => NotFound(),
        DraftReviewStatus.InvalidGeometry => Problem(
            detail: "Geometry must be valid GeoJSON matching the draft's feature type (LineString for roads, Polygon for buildings).",
            statusCode: 400),
        DraftReviewStatus.RulesNotMet => Problem(
            detail: "Draft does not meet the configured road rules (minimum length / minimum confidence).",
            statusCode: 422),
        DraftReviewStatus.AlreadyReviewed => Problem(detail: $"Draft is not pending.", statusCode: 409),
        DraftReviewStatus.Forbidden => Forbid(),
        _ => UnknownReviewStatus(result.Status),
    };

    /// <summary>
    /// A future <see cref="DraftReviewStatus"/> value should surface as a
    /// logged server error, not a misleading 403.
    /// </summary>
    private IActionResult UnknownReviewStatus(DraftReviewStatus status)
    {
        logger.LogError("Unhandled DraftReviewStatus '{Status}' returned by the draft-features service", status);
        return Problem(detail: "An internal error occurred.", statusCode: 500);
    }
}